feat: SM-74 스타일드 컴포넌트와 인터페이스 파일 분리
